fix drake initialization and add QOL improvements

+++ scripts/visualize_pendulum.py
@@ -26,7 +26,7 @@ def _policy(state: TrajOptState, env: PendulumTrajOptProblem) -> TrajOptAction:
     k = 100.0
     u = -k * theta_dot * (current_energy - desired_energy)
     u = np.clip(u, -2.0, 2.0)  # to make this nontrivial
-    return np.array([u], dtype=np.float32)
+    return np.array([u], dtype=np.float64)
 
 
 def _main(max_horizon: int, outdir: Path, fps: int) -> None:
