+++ src/mlrp_course/trajopt/envs/double_integrator.py
@@ -64,7 +64,7 @@ class DoubleIntegratorProblem(TrajOptProblem):
 
     @property
     def initial_state(self) -> TrajOptState:
-        return np.array([-1.0, 0.0], dtype=np.float32)
+        return np.array([-1.0, 0.0], dtype=np.float64)
 
     def get_next_state(
         self, state: TrajOptState, action: TrajOptAction
@@ -103,9 +103,9 @@ class DoubleIntegratorProblem(TrajOptProblem):
 
     @staticmethod
     def _get_traj_cost(
-        xs: NDArray[np.float32],
-        x_dots: NDArray[np.float32],
-        actions: NDArray[np.float32],
+        xs: NDArray[np.float64],
+        x_dots: NDArray[np.float64],
+        actions: NDArray[np.float64],
         x_cost_weight: float,
         x_dot_cost_weight: float,
         torque_cost_weight: float,
@@ -140,7 +140,7 @@ class JaxDoubleIntegratorProblem(DoubleIntegratorProblem):
 
     @property
     def initial_state(self) -> TrajOptState:
-        return jnp.array(super().initial_state, dtype=jnp.float32)
+        return jnp.array(super().initial_state, dtype=jnp.float64)
 
     @staticmethod
     @jax.jit
@@ -156,14 +156,14 @@ class JaxDoubleIntegratorProblem(DoubleIntegratorProblem):
         next_x_dot = x_dot + u * dt
         next_x = x + next_x_dot * dt
 
-        return jnp.array([next_x, next_x_dot], dtype=jnp.float32)
+        return jnp.array([next_x, next_x_dot], dtype=jnp.float64)
 
     @staticmethod
     @jax.jit
     def _get_traj_cost(
-        xs: NDArray[jnp.float32],
-        x_dots: NDArray[jnp.float32],
-        actions: NDArray[jnp.float32],
+        xs: NDArray[jnp.float64],
+        x_dots: NDArray[jnp.float64],
+        actions: NDArray[jnp.float64],
         x_cost_weight: float,
         x_dot_cost_weight: float,
         torque_cost_weight: float,
